2023/10/24

+++ Common/MathUtils.java
@@ -14,7 +14,7 @@ public class MathUtils {
         return rtn;
     }
 
-    public static Point ClacQuadratic (int a, int b, int c, int x) {
+    public static Point ClacQuadratic (int a, int b, int c, double x) {
         double moveY = a * Math.pow(x, 2) + b * x + c;
 
         Point rtn = new Point();
